requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
idna==2.7
jsmin==2.2.2
psycopg2==2.8.5
pytz==2020.1
requests==2.19.1
sqlparse==0.3.1
urllib3==1.23
whitenoise==5.1.0
wincertstore==0.2
